pyproject.toml
[project]
name = "memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "groq>=1.2.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.4.1",
    "uuid>=1.30",
    "uvicorn>=0.45.0",
]
